refactor(e2e): share readMigratedState across the migration specs
Three migration specs carried byte-identical copies of the SUP_OPS snapshot
read, the third added by the #9770 test. Moves it to the helper module the
suite already uses for seedLegacyDatabase, generic over the slices each test
asserts on so the narrow return types survive.

Also updates "should handle migration error gracefully", which pinned the old
hardcoded error string and a single-button dialog — both changed by the
preceding two commits.
